Move pull request job into webhooks
